Exclude certain ACM namespaces from network policy
The require-namespace-network-policies constraint requires that every
namespace in the cluster have a NetworkPolicy object defined.

Certain ACM products don't include such a NetworkPolicy, and are thus
broken by this Constraint if it is applied in its current state.

To solve this problem, this PR adds the following namespaces as
`excludedNamespaces` in the `match` section of the aforementioned
Constraint:

  - kube-system
  - config-management-system
  - config-management-monitoring
  - resource-group-system